showPrivacyPolicyToast
