+++ ui/pages/routes/routes.component.js
@@ -147,7 +147,7 @@ import {
   setTheme,
   showOnboardingHeader,
 } from './isolated';
-import { ToastMaster, updateNewPrivacyPolicyToastDate } from './toast-master';
+import { ToastMaster } from './toast-master';
 
 export default class Routes extends Component {
   static propTypes = {
@@ -214,12 +214,8 @@ export default class Routes extends Component {
     totalUnapprovedConfirmationCount: PropTypes.number.isRequired,
     currentExtensionPopupId: PropTypes.number,
     useRequestQueue: PropTypes.bool,
-    showPrivacyPolicyToast: PropTypes.bool.isRequired,
-    newPrivacyPolicyToastShownDate: PropTypes.number,
     setSurveyLinkLastClickedOrClosed: PropTypes.func.isRequired,
-    setNewPrivacyPolicyToastShownDate: PropTypes.func.isRequired,
     clearEditedNetwork: PropTypes.func.isRequired,
-    setNewPrivacyPolicyToastClickedOrClosed: PropTypes.func.isRequired,
     ///: BEGIN:ONLY_INCLUDE_IF(keyring-snaps)
     isShowKeyringSnapRemovalResultModal: PropTypes.bool.isRequired,
     hideShowKeyringSnapRemovalResultModal: PropTypes.func.isRequired,
@@ -231,10 +227,6 @@ export default class Routes extends Component {
     t: PropTypes.func,
     metricsEvent: PropTypes.func,
   };
-
-  componentDidMount() {
-    updateNewPrivacyPolicyToastDate(this.props);
-  }
 
   componentDidUpdate(prevProps) {
     const {
