+++ ui/pages/routes/toast-master.test.ts
@@ -1,5 +1,9 @@
+import { PRIVACY_POLICY_DATE } from '../../helpers/constants/privacy-policy';
 import { SURVEY_DATE, SURVEY_GMT } from '../../helpers/constants/survey';
-import { getShowSurveyToast } from './toast-master';
+import {
+  getShowSurveyToast,
+  getShowPrivacyPolicyToast,
+} from './toast-master-selectors';
 
 describe('#getShowSurveyToast', () => {
   const realDateNow = Date.now;
@@ -50,5 +54,141 @@ describe('#getShowSurveyToast', () => {
       },
     });
     expect(result).toStrictEqual(false);
+  });
+});
+
+describe('#getShowPrivacyPolicyToast', () => {
+  let dateNowSpy: jest.SpyInstance;
+
+  describe('mock one day after', () => {
+    beforeEach(() => {
+      const dayAfterPolicyDate = new Date(PRIVACY_POLICY_DATE);
+      dayAfterPolicyDate.setDate(dayAfterPolicyDate.getDate() + 1);
+
+      dateNowSpy = jest
+        .spyOn(Date, 'now')
+        .mockReturnValue(dayAfterPolicyDate.getTime());
+    });
+
+    afterEach(() => {
+      dateNowSpy.mockRestore();
+    });
+
+    it('shows the privacy policy toast when not yet seen, on or after the policy date, and onboardingDate is before the policy date', () => {
+      const result = getShowPrivacyPolicyToast({
+        metamask: {
+          newPrivacyPolicyToastClickedOrClosed: null,
+          onboardingDate: new Date(PRIVACY_POLICY_DATE).setDate(
+            new Date(PRIVACY_POLICY_DATE).getDate() - 2,
+          ),
+        },
+      });
+      expect(result).toBe(true);
+    });
+
+    it('does not show the privacy policy toast when seen, even if on or after the policy date and onboardingDate is before the policy date', () => {
+      const result = getShowPrivacyPolicyToast({
+        metamask: {
+          newPrivacyPolicyToastClickedOrClosed: true,
+          onboardingDate: new Date(PRIVACY_POLICY_DATE).setDate(
+            new Date(PRIVACY_POLICY_DATE).getDate() - 2,
+          ),
+        },
+      });
+      expect(result).toBe(false);
+    });
+
+    it('shows the privacy policy toast when not yet seen, on or after the policy date, and onboardingDate is not set', () => {
+      const result = getShowPrivacyPolicyToast({
+        metamask: {
+          newPrivacyPolicyToastClickedOrClosed: null,
+          onboardingDate: null,
+        },
+      });
+      expect(result).toBe(true);
+    });
+  });
+
+  describe('mock same day', () => {
+    beforeEach(() => {
+      dateNowSpy = jest
+        .spyOn(Date, 'now')
+        .mockReturnValue(new Date(PRIVACY_POLICY_DATE).getTime());
+    });
+
+    afterEach(() => {
+      dateNowSpy.mockRestore();
+    });
+
+    it('shows the privacy policy toast when not yet seen, on or after the policy date, and onboardingDate is before the policy date', () => {
+      const result = getShowPrivacyPolicyToast({
+        metamask: {
+          newPrivacyPolicyToastClickedOrClosed: null,
+          onboardingDate: new Date(PRIVACY_POLICY_DATE).setDate(
+            new Date(PRIVACY_POLICY_DATE).getDate() - 2,
+          ),
+        },
+      });
+      expect(result).toBe(true);
+    });
+
+    it('does not show the privacy policy toast when seen, even if on or after the policy date and onboardingDate is before the policy date', () => {
+      const result = getShowPrivacyPolicyToast({
+        metamask: {
+          newPrivacyPolicyToastClickedOrClosed: true,
+          onboardingDate: new Date(PRIVACY_POLICY_DATE).setDate(
+            new Date(PRIVACY_POLICY_DATE).getDate() - 2,
+          ),
+        },
+      });
+      expect(result).toBe(false);
+    });
+
+    it('shows the privacy policy toast when not yet seen, on or after the policy date, and onboardingDate is not set', () => {
+      const result = getShowPrivacyPolicyToast({
+        metamask: {
+          newPrivacyPolicyToastClickedOrClosed: null,
+          onboardingDate: null,
+        },
+      });
+      expect(result).toBe(true);
+    });
+  });
+
+  describe('mock day before', () => {
+    beforeEach(() => {
+      const dayBeforePolicyDate = new Date(PRIVACY_POLICY_DATE);
+      dayBeforePolicyDate.setDate(dayBeforePolicyDate.getDate() - 1);
+
+      dateNowSpy = jest
+        .spyOn(Date, 'now')
+        .mockReturnValue(dayBeforePolicyDate.getTime());
+    });
+
+    afterEach(() => {
+      dateNowSpy.mockRestore();
+    });
+
+    it('does not show the privacy policy toast before the policy date', () => {
+      const result = getShowPrivacyPolicyToast({
+        metamask: {
+          newPrivacyPolicyToastClickedOrClosed: null,
+          onboardingDate: new Date(PRIVACY_POLICY_DATE).setDate(
+            new Date(PRIVACY_POLICY_DATE).getDate() - 2,
+          ),
+        },
+      });
+      expect(result).toBe(false);
+    });
+
+    it('does not show the privacy policy toast before the policy date even if onboardingDate is not set', () => {
+      const result = getShowPrivacyPolicyToast({
+        metamask: {
+          newPrivacyPolicyToastClickedOrClosed: null,
+          onboardingDate: null,
+        },
+      });
+      expect(result).toBe(false);
+    });
   });
 });
