+++ ui/store/actions.ts
@@ -4285,12 +4285,6 @@ export function setSurveyLinkLastClickedOrClosed(time: number) {
   };
 }
 
-export function setNewPrivacyPolicyToastClickedOrClosed() {
-  return async () => {
-    await submitRequestToBackground('setNewPrivacyPolicyToastClickedOrClosed');
-  };
-}
-
 export function setLastViewedUserSurvey(id: number) {
   return async () => {
     await submitRequestToBackground('setLastViewedUserSurvey', [id]);
@@ -4300,14 +4294,6 @@ export function setLastViewedUserSurvey(id: number) {
 export function setOnboardingDate() {
   return async () => {
     await submitRequestToBackground('setOnboardingDate');
-  };
-}
-
-export function setNewPrivacyPolicyToastShownDate(time: number) {
-  return async () => {
-    await submitRequestToBackground('setNewPrivacyPolicyToastShownDate', [
-      time,
-    ]);
   };
 }
